Ensure that editable wheels also get lexers/parsers built (#6)
For the sake of compatibility with some plugins, setuptools changed
the behavior of the internal `editable_wheel` command not to run
the `build` command, but to run its subcommands only. However, this
broke ANTLeRinator, as it relied on `build` to run `build_antlr` to
build any lexers/parsers. Now, `editable_wheel` is patched to
ensure that `build_antlr` is called.